fix: use static params in file-system based router for ssg
